another commit

+++ examples/Equal.c++
@@ -18,8 +18,8 @@ void test1 () {
     assert(!equal(a, a + 3, b));}
 
 void test2 () {
-    const int a[] = {2, 3, 4};
-    const int b[] = {0, 2, 3, 4, 0};
+    const int  a[] = {2, 3, 4};
+    const long b[] = {0, 2, 3, 4, 0};
     assert(equal(a, a + 3, b + 1));}
 
 void test3 () {
